[refactor] 클래스 명칭 변경
