+++ src/main/java/bowling/controller/BowlingGameController.java
@@ -1,6 +1,6 @@
 package bowling.controller;
 
-import bowling.domain.bowlinggame.BowlingGameBoard;
+import bowling.domain.frame.Frames;
 import bowling.domain.user.User;
 import bowling.view.InputView;
 import bowling.view.ResultView;
@@ -11,13 +11,13 @@ public class BowlingGameController {
         ResultView.printAskUserNameMessage();
         User user = new User(InputView.askUserName());
 
-        BowlingGameBoard bowlingGameBoard = new BowlingGameBoard();
-        ResultView.printScoreBoard(user.makeDTO(), bowlingGameBoard.exportStatus().makeDTO());
+        Frames frames = new Frames();
+        ResultView.printScoreBoard(user.makeDTO(), frames.exportResult().makeDTO());
 
-        while (!bowlingGameBoard.isEnd()) {
-            ResultView.printCurrentFrameNumber(bowlingGameBoard.exportStatus().makeDTO());
-            bowlingGameBoard.record(InputView.askScore());
-            ResultView.printScoreBoard(user.makeDTO(), bowlingGameBoard.exportStatus().makeDTO());
+        while (!frames.isEnd()) {
+            ResultView.printCurrentFrameNumber(frames.exportResult().makeDTO());
+            frames.record(InputView.askScore());
+            ResultView.printScoreBoard(user.makeDTO(), frames.exportResult().makeDTO());
         }
 
     }
