+++ src/main/java/bowling/dto/BowlingGameResultsDTO.java
@@ -3,12 +3,12 @@ package bowling.dto;
 import java.util.Collections;
 import java.util.List;
 
-public class BowlingStatusDTO {
+public class BowlingGameResultsDTO {
 
     private final int currentFrame;
     private final List<String> frames;
 
-    public BowlingStatusDTO(int currentFrame, List<String> frames) {
+    public BowlingGameResultsDTO(int currentFrame, List<String> frames) {
         this.currentFrame = currentFrame;
         this.frames = frames;
     }
